+++ content/Algorithms/DFM/DFM.tpp
@@ -10,7 +10,7 @@ void DFM<T>::initialize(std::vector<Data<T>>& R, DataSet<T>& O,
 }
 
 template<typename T>
-std::vector<float> DFM<T>::execute() {
+std::vector<double> DFM<T>::execute() {
     // TODO: Implementar el algoritmo principal DFM
     answer[0] = execute_dunn_star();
     answer[1] = execute_db_star();
@@ -24,23 +24,23 @@ std::vector<float> DFM<T>::execute() {
 }
 
 template<typename T>
-float DFM<T>::execute_dunn_star() {
+double DFM<T>::execute_dunn_star() {
     // TODO: Implementar el índice Dunn*
     return 0.0f;
 }
 
 template<typename T>
-float DFM<T>::execute_db_star() {
+double DFM<T>::execute_db_star() {
     // TODO: Implementar el índice Davies-Bouldin*
     return 0.0f;
 }
 
 template<typename T>
-float DFM<T>::execute_sil_star() {
+double DFM<T>::execute_sil_star() {
     //Calculo de los medoides mas cercanos
     std::vector<Data<T>> NearestMedoid;
     for(int i=0; i<C.size(); i++){
-        float min_distance = std::numeric_limits<float>::max();
+        double min_distance = std::numeric_limits<double>::max();
         int nearest_medoid_index = -1;
         for(int j=0; j<C.size(); j++){
             if(i != j){
@@ -54,12 +54,12 @@ float DFM<T>::execute_sil_star() {
     }
 
     //Calculo de los coeficientes de silueta
-    float SumatoriaFinal = 0;
-    float resultParcial;
+    double SumatoriaFinal = 0;
+    double resultParcial;
     for(int i = 0; i<C.size();i++){
-        float resultParcial = 0;
-        float resNum = 0;
-        float resDen = 0;
+        double resultParcial = 0;
+        double resNum = 0;
+        double resDen = 0;
         for(int j=0; j<C[i].size(); j++){
             resNum += distance_div(C[i].getData(j), NearestMedoid[i]);
         }
@@ -73,25 +73,25 @@ float DFM<T>::execute_sil_star() {
 }
 
 template<typename T>
-float DFM<T>::execute_sigma_sim() {
+double DFM<T>::execute_sigma_sim() {
     // TODO: Implementar la medida sigma de similitud
     return 0.0f;
 }
 
 template<typename T>
-float DFM<T>::execute_sigma_div() {
+double DFM<T>::execute_sigma_div() {
     // TODO: Implementar la medida sigma de diversidad
     return 0.0f;
 }
 
 template<typename T>
-float DFM<T>::execute_u_sim() {
+double DFM<T>::execute_u_sim() {
     // TODO: Implementar la medida U de similitud
     return 0.0f;
 }
 
 template<typename T>
-float DFM<T>::execute_u_div() {
+double DFM<T>::execute_u_div() {
     // TODO: Implementar la medida U de diversidad
     return 0.0f;
 }
